+++ .gitignore
@@ -4,3 +4,4 @@ __pycache__/config.cpython-310.pyc
 __pycache__/loaders.cpython-310.pyc
 data/__pycache__/loaders.cpython-310.pyc
 game/__pycache__/state.cpython-310.pyc
+game/__pycache__/help.cpython-310.pyc
